+++ pkgs/test/CHANGELOG.md
@@ -11,7 +11,7 @@
   all tests with OS `'windows'` would previously still run browser tests on
   windows, but will now skip all tests including browser tests.
 * Use a DevTools URL instead of a defunct observatory URL.
-* Add flag `--shard-by` to control sharding strategy.
+* Add flag `--shard-by-file` to control sharding strategy.
 
 ## 1.31.1